move tests to folder
