+++ webui/react/src/e2e/tests/navigation.spec.ts
@@ -1,6 +1,6 @@
 import { expect } from '@playwright/test';
 
-import { test } from './global-fixtures';
+import { test } from '../fixtures/global-fixtures';
 
 test.describe('Navigation', () => {
   const USERNAME = process.env.PW_USER_NAME ?? '';
